chat-bot-comm/requirements.txt
aiohttp==3.8.1
pytest==7.1.2
pytest-aiohttp==1.0.4
pytest-asyncio==0.19.0
pytest-env==0.6.2
marshmallow==3.20.1
marshmallow-dataclass==8.6.0
PyYAML==6.0.1
aiohttp_cors==0.7.0
aiohttp-apispec==2.2.3
pika==1.3.2
aio-pika==9.3.1
python-dotenv==1.0.1
redis==5.0.4